Integrate TaskManager and JSONParser
